ent-v3/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-jose==3.3.0
python-dotenv==1.0.0
minio==7.1.17
cassandra-driver==3.28.0
pydantic==2.4.2
requests==2.31.0
python-keycloak==3.3.0